pyproject.toml
[project]
name = "market-analyst-agent"
version = "0.1.0"
description = "Institutional Market Analyst Agent - Demo for Engineering the Agentic Stack blog series"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-anthropic>=0.3.0",
    "langgraph-checkpoint-postgres>=2.0.0",
    "yfinance>=0.2.50",
    "tavily-python>=0.5.0",
    "pydantic>=2.0.0",
    "redis>=5.0.0",
    "python-dotenv>=1.0.0",
    "psycopg[binary]>=3.2.0",
]

[project.scripts]
market-analyst = "market_analyst.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/market_analyst"]
